+++ packages/spry/example/main.dart
@@ -4,6 +4,9 @@ final app = Application.late();
 
 void main(List<String> args) async {
   app.get('hello', (request) => 'Hello, Spry!');
+  app.get('test', (request) {
+    return request.fetch('/hello');
+  });
 
   await app.run(port: 3000);